move risk-advisor service to a separate package, minor refactoring
