+++ riskadvisor/server.go
@@ -1,28 +1,14 @@
-package main
+package riskadvisor
 
 import (
 	"bytes"
 	"encoding/json"
-	"fmt"
 	"io/ioutil"
-	"log"
 	"net/http"
 
 	"github.com/emicklei/go-restful"
-	flag "github.com/spf13/pflag"
 	"k8s.io/kubernetes/pkg/api"
 )
-
-type AdviceRequest struct {
-	Id  int     `json:"id"`
-	Pod api.Pod `json:"pod"`
-}
-
-type AdviceStatus struct {
-	Id     int         `json:"id"`
-	Status string      `json:"status"`
-	Result api.Binding `json:"result"`
-}
 
 type AdviceService struct {
 	proxyUrl string
@@ -38,26 +24,26 @@ func (a *AdviceService) sendAdviceRequest(request *restful.Request, response *re
 	err := request.ReadEntity(&pod)
 	if err == nil {
 		ar := AdviceRequest{thisId, *pod}
-		ar_json, err := json.Marshal(ar)
+		arJSON, err := json.Marshal(ar)
 		if err != nil {
 			response.WriteError(http.StatusInternalServerError, err)
 			return
 		}
 
-		resp, err := http.Post(a.proxyUrl, "application/json", bytes.NewReader(ar_json))
+		resp, err := http.Post(a.proxyUrl, "application/json", bytes.NewReader(arJSON))
 		if err != nil {
 			response.WriteError(http.StatusNotFound, err)
 			return
 		}
 
-		binding_json, err := ioutil.ReadAll(resp.Body)
+		bindingJSON, err := ioutil.ReadAll(resp.Body)
 		if err != nil {
 			response.WriteError(http.StatusInternalServerError, err)
 			return
 		}
 
 		binding := api.Binding{}
-		err = json.Unmarshal(binding_json, &binding)
+		err = json.Unmarshal(bindingJSON, &binding)
 		if err != nil {
 			response.WriteError(http.StatusExpectationFailed, err)
 			return
@@ -70,7 +56,8 @@ func (a *AdviceService) sendAdviceRequest(request *restful.Request, response *re
 	}
 }
 
-func main() {
+// Comment it for now, will need to be refactored
+/*func main() {
 	var port *int = flag.Int("port", 11111, "Port to listen on")
 	var proxyUrl *string = flag.String("proxy-url", "http://localhost:9998/", "URL of proxy server")
 	flag.Parse()
@@ -82,7 +69,7 @@ func main() {
 		Produces(restful.MIME_JSON)
 
 	aws.Route(aws.POST("").To(as.sendAdviceRequest).
-		// Documentation
+	// Documentation
 		Doc("Post a request for advice").
 		Reads(api.Pod{}).
 		Returns(200, "OK", AdviceStatus{}))
@@ -91,4 +78,4 @@ func main() {
 
 	log.Printf("start listening on localhost:%d", *port)
 	log.Fatal(http.ListenAndServe(fmt.Sprintf(":%d", *port), nil))
-}
+}*/
